trying out a different move to player
