+++ Assets/Scripts/Monster/MonsterSight.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: f2b09759a1f50aa4eae78f59c7f5b5d3
+timeCreated: 1524138527
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
